Rename ContingenciesProviderMock and ContingenciesProviderFactoryMock to EmptyContingencyListProvider and EmptyContingencyListProviderFactory
