+++ contingency-api/src/main/java/eu/itesla_project/contingency/EmptyContingencyListProvider.java
@@ -4,10 +4,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package eu.itesla_project.contingency.mock;
+package eu.itesla_project.contingency;
 
-import eu.itesla_project.contingency.ContingenciesProvider;
-import eu.itesla_project.contingency.Contingency;
 import eu.itesla_project.iidm.network.Network;
 
 import java.util.Collections;
@@ -16,7 +14,7 @@ import java.util.List;
 /**
  * @author Mathieu Bague <mathieu.bague at rte-france.com>
  */
-public class ContingenciesProviderMock implements ContingenciesProvider {
+public class EmptyContingencyListProvider implements ContingenciesProvider {
 
     @Override
     public List<Contingency> getContingencies(Network network) {
