+++ contingency-api/src/test/java/eu/itesla_project/contingency/EmptyContingencyListProviderTest.java
@@ -4,24 +4,22 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package eu.itesla_project.contingency.mock;
+package eu.itesla_project.contingency;
 
-import eu.itesla_project.contingency.ContingenciesProvider;
-import eu.itesla_project.contingency.ContingenciesProviderFactory;
 import org.junit.Assert;
 import org.junit.Test;
 
 /**
  * @author Mathieu Bague <mathieu.bague at rte-france.com>
  */
-public class ContingenciesProviderTest {
+public class EmptyContingencyListProviderTest {
 
     @Test
     public void test() {
-        ContingenciesProviderFactory factory = new ContingenciesProviderFactoryMock();
+        ContingenciesProviderFactory factory = new EmptyContingencyListProviderFactory();
         ContingenciesProvider provider = factory.create();
 
-        Assert.assertTrue(provider instanceof ContingenciesProviderMock);
+        Assert.assertTrue(provider instanceof EmptyContingencyListProvider);
         Assert.assertEquals(0, provider.getContingencies(null).size());
     }
 }
